# Optimize by avoiding pattern matching