--- a/lib/fusuma/plugin/filters/libinput_device_filter.rb
+++ b/lib/fusuma/plugin/filters/libinput_device_filter.rb
@@ -26,7 +26,7 @@ module Fusuma
             return false
           end
           device_id = record.to_s.match(/\S*/, 1).to_s
-          keep_device.all_ids.include?(device_id)
+          keep_device.all.map(&:id).include?(device_id)
         end
 
         def keep_device
@@ -58,13 +58,13 @@ module Fusuma
 
           # remove cache for reloading new devices
           def reset
-            @all_ids = nil
+            @all = nil
             Device.reset
           end
 
           # @return [Array]
-          def all_ids
-            @all_ids ||= if @name_patterns.empty?
+          def all
+            @all ||= if @name_patterns.empty?
               Device.available
             else
               Device.all.select do |device|
@@ -72,7 +72,7 @@ module Fusuma
               end
             end.tap do |devices|
               print_not_found_messages if devices.empty?
-            end.map(&:id)
+            end
           end
 
           def print_not_found_messages