Apparently ruby's string matching is quite slow. This makes the `String.underscore` function slow, so the first commit avoids calling it in the hot code path. The second commit speeds up the pattern-matching when filtering by device (this is instead of the `@keep_all` approach I suggested before).

Benchmark: I ran `20000.times { pipeline }` with both versions, feeding them touchpad events that I had recorded for a couple of hours. The results show a 18% speedup

| version | time |
|----------|--------------------|
| b68ab603 | 5.024 s ±  0.058 s |
| this PR  | 4.256 s ±  0.035 s |