@iberianpig 2023-07-20
Please fix failed the spec. https://github.com/iberianpig/fusuma/actions/runs/5574219886/jobs/10182543294?pr=304